Fix bug of z-index after switching pages
