+++ package.json
@@ -14,12 +14,15 @@
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.14.14",
     "@mui/system": "^5.14.14",
+    "@types/nodemailer": "^6.4.13",
     "clsx": "^2.0.0",
     "framer-motion": "^10.16.4",
     "next": "13.5.4",
+    "nodemailer": "^6.9.7",
     "react": "^18",
     "react-bootstrap": "^2.9.0",
     "react-dom": "^18",
+    "react-hook-form": "^7.47.0",
     "react-icons": "^4.11.0",
     "react-iframe": "^1.8.5",
     "tesseract.js": "^5.0.2"
